Enable GST_STATE_PAUSED again
Guard if it can't be paused when VIDIOC_TRY_DECODER_CMD was thrown
from Chromium, because usually it is not playing nor ready state yet.
(before playing video stream)

Basically set_decoder_cmd_state should return -1 on error, but it was
treated as if VIDIOC_TRY_DECODER_CMD was not supported in Chromium's
IsDecoderCmdSupported(), so just return 0 here.

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>